move scripts into scripts dir and update paths accordingly
