+++ scripts/set-version-bump.js
@@ -1,7 +1,18 @@
+/**
+ * For patch releases, use `npm run set-version-bump` script.
+ * For minor or major releases, use `node run ./scripts/set-version-bump.js minor`.
+ * - Then run the remaining steps from the npm script:
+ * - `npm i && rm -rf ./build && npm run build && composer run i18n`
+ */
+
 const fs = require( 'fs' );
+const path = require( 'path' );
 const semver = require( 'semver' );
-const packagefile = './package.json';
-const pluginfile = './bluehost-wordpress-plugin.php';
+const packagefile = path.resolve( __dirname, '../package.json' );
+const pluginfile = path.resolve(
+	__dirname,
+	'../bluehost-wordpress-plugin.php'
+);
 
 if ( fs.existsSync( packagefile ) && fs.existsSync( pluginfile ) ) {
 	const packageData = require( packagefile );
@@ -29,4 +40,8 @@ if ( fs.existsSync( packagefile ) && fs.existsSync( pluginfile ) ) {
 	} );
 
 	console.log( 'Version updated', currentVersion, '=>', newVersion );
+} else {
+	console.log(
+		'Version update error: package.json or bluehost-wordpress-plugin.php not found'
+	);
 }
